+++ src/EmbDB/EmbDB/DB/BTreePlus/BaseLeafNodeSetCompressor.h
@@ -79,7 +79,7 @@ namespace embDB
 			virtual bool add(const TKeyMemSet& vecKeys)
 			{
 
-				for (uint32 i = 0, sz = vecKeys.size(); i < sz; ++i)
+				for (uint32 i = 0, sz = (uint32)vecKeys.size(); i < sz; ++i)
 				{
 					m_KeyEncoder.AddSymbol(m_nCount, m_nCount + i, vecKeys[i], vecKeys);
 					m_nCount++;
@@ -89,7 +89,7 @@ namespace embDB
 			virtual bool recalc(const TKeyMemSet& vecKeys)
 			{
 				clear();
-				for (uint32 i = 0, sz = vecKeys.size(); i < sz; ++i)
+				for (uint32 i = 0, sz = (uint32)vecKeys.size(); i < sz; ++i)
 				{
 					m_nCount += 1;
 					m_KeyEncoder.AddSymbol(m_nCount, i, vecKeys[i], vecKeys);
@@ -144,7 +144,7 @@ namespace embDB
 			{
 				m_nCount = 0;
 				m_KeyEncoder.clear();
-				for (uint32 i = 0, sz = vecKeys.size(); i < sz; ++i)
+				for (uint32 i = 0, sz = (uint32)vecKeys.size(); i < sz; ++i)
 				{
 					m_nCount++;
 					m_KeyEncoder.AddSymbol(m_nCount, i, vecKeys[i], vecKeys);
